Initial ContributorMonthList tests
