+++ public/models/contribution-month/test/test.js
@@ -1,3 +1,4 @@
 require('./contribution-month_test');
-require('./monthly-os-project_test')
-require('./monthly-client-project_test')
+require('./monthly-os-project_test');
+require('./monthly-client-project_test');
+require('./contribution-month-list_test');
